Raise time limit before reconnecting for no msgs
Previously after five minutes if it hadn't received any messages piaware
would kill and restart faup1090 and reconnect to it in order to get it
to reconnect to dump1090 or modesmixer or whatnot, just in case there
was a problem with the connection.

FlightAware user "fill" pointed out that in off hours it will be common for
many receivers to not see any messages for more than five minutes.

The timeout is raised to one hour and the log message now better explains
what's going on:

"no new messages received in $secondsSinceLast seconds, it might just be that
there haven't been any aircraft nearby but I'm going to possibly restart
faup1090 and definitely reconnect, just in case there's a problem with the
current connection..."