PiAware logging improvements
* Now logs the number of ADS-B messages received in the reporting interval
(5 minutes) as well as cumulative, which makes it easier to see how many
messages you are getting without having to do arithmetic in your head.

* When piaware is asked by the system to shutdown, its log messages now says
that in a more clear way.

* cryptic "not-connected-yet" logged in some cases as a message source
replaced by the hopefully less cryptic
"(not currently connected to an adsb source)", which may be true when
we have been connected but are no longer.