Fuzz reconnect interval connecting to FlightAware
If the adept server shuts down and drops all the clients, they all delay 60 seconds and try to reconnect, which is a lot of connections all at once, which is kind of staggering to the infrastructure.  By fuzzing the delay between 60 and 120 seconds (instead of a fixed 60 seconds) it will at least somewhat spread out the connection attempts, though they will still be frequent.

If the server is down more than a couple minutes then when it comes back up the connections should be pretty randomized with 100% of the remotes making connection attempts within two minutes.