Report package version separately. Don't report empty vars.
The package version may not reflect what's actually running (for
example, my dev system claims to be 1.21, but that's actually a
removed-but-not-purged package from long ago), so report it
separately rather than clobbering the hardcoded values. The
hardcoded values should "always" be correct now that we have a
check in the packaging process.