Change the default port for mlat output to 30104.
This is so we don't automatically connect to a local dump1090
that may not understand mlat - user intervention is needed in this
case.

The default piaware config will have a matching setup where there
is a mlat-aware dump1090 listening on 30104; I am also going to
update the mlat-aware dump1090-mutability to listen on 30104.

See github issue #24 for some background.