Don't install sources.list/FA key in run-apt-get. Try upgrades when requested even if the OS doesn't seem supported.
Now we assume that a suitable piaware-repository package is installed,
as this should be the case for both piaware sdcard images and for
FA-built packages installed from the FA repository.

For self-built packages there will probably be no repository installed
and the update won't find anything, but that's expected - if it's self-
built then presumably it's a non-official build and upgrading from
the FA repo doesn't make sense anyway.